+++ platformio/debug/config/blackmagic.py
@@ -14,7 +14,7 @@
 
 from platformio.debug.config.base import DebugConfigBase
 from platformio.debug.exception import DebugInvalidOptionsError
-from platformio.device.finder import find_serial_port, is_pattern_port
+from platformio.device.finder import SerialPortFinder, is_pattern_port
 
 
 class BlackmagicDebugConfig(DebugConfigBase):
@@ -56,12 +56,11 @@ set language auto
         initial_port = DebugConfigBase.port.fget(self)
         if initial_port and not is_pattern_port(initial_port):
             return initial_port
-        port = find_serial_port(
-            initial_port,
+        port = SerialPortFinder(
             board_config=self.board_config,
             upload_protocol=self.tool_name,
             prefer_gdb_port=True,
-        )
+        ).find(initial_port)
         if port:
             return port
         raise DebugInvalidOptionsError(