+++ HISTORY.rst
@@ -16,6 +16,7 @@ PlatformIO Core 6
 6.1.4 (2022-??-??)
 ~~~~~~~~~~~~~~~~~~
 
+* Improved device port finder when using dual channel UART converter (`issue #4367 <https://github.com/platformio/platformio-core/issues/4367>`_)
 * Improved project dependency resolving when using the `pio project init --ide <https://docs.platformio.org/en/latest/core/userguide/project/cmd_init.html>`__ command
 * Fixed an issue when escaping macros/defines for IDE integration (`issue #4360 <https://github.com/platformio/platformio-core/issues/4360>`_)